All your transformers are belong to us.
All transformers are implemented: reader,writer,state,fail,nondet.

All of the interactions between transformers are implemented.

Everything is tested.

I added a new piece of syntax to do-notation that lets you do:

(do
  x ← (return 1)
  y ≔ (+ x 1)
  z ← (return y)
  (return z))

The syntax `do x ≔ e . body` expands to `(let ([x e]) (do . body)`.

In particular, the summarizing behavior of nondeterminism on top of
state, which is what results in flow insensitivity for that piece of
state, is implemented and tested. Use the FinMapO monoid to build finite
maps which can be summarized.

(for example, to build a map from integers to powersets, where the
powersets are unioned when two finite maps need to be summarized, just
use `FinMapO PowersetO`)