Restrict closures to free variables; breaks GC.
This commit changes closure creation to restrict the environment to
the domain of free variables in the term.  Doing so breaks tests for
things involving GC, which I think shows there's an overall brokeness
to the current GC implementation.

I have an alternate implementation that is working fine for concrete
evaluation.  I'll try to scale it up to PDCFA and push if it fixes
this problem.