PDCFA in eval/apply form.
This is PDCFA applied to the eval/apply variant of the evaluator.
This is a first step toward recreating Big CFA2 from Vardoulakis's
dissertation, which cache's on apply, not eval as we do.

This set-up still caches on eval and has an outer iterator, but is
organized with mutually (open) recursive eval and apply functions.