Still stuck with issue #78.
I've tried my hardest to get the appendix to be included in the paper,
but it doesn't seem to work.  In the appendix directory there is a
minimal tex document that is able to \input the appendix.  I've
included the relevant definitions and packages in style.tex.  But for
some reason it crashes when I do \input from the Scribble generated
main.tex file complaining about missing braces.