Fix typos
1. The cache name is ‘$in’, not ‘in’.
2. The escape character in Scribble is ‘@’, not ‘$’.
3. The Figure in question is 10, not 11.

(It’s kind of funny that 2 made into the paper. At first I was wondering
if you had used Scribble, but at that point I was sure, even before
looking at the sources 😛)

There’s actually a fourth bug: on § 10, there’s a psychologically bad
line break [1] before the reference [King 1976]. Unfortunately, I don’t
know how to fix this in Scribble. A hint might be the name of the
citation form, ‘~cite’, which seems to indicate that it inserts a
non-breaking space before the citation (in TeX, a non-breaking space is
spelled ‘~’, hence my intuition). If this is right, then the solution is
simple: everywhere this form is used, remove the space right before it,
for example, ‘closely@~cite{dvanhorn:King1976Symbolic}’.

This was an interesting and fun paper to read. Thank you for your work.

[1]: The TeXbook. Donald Ervin Knuth.